REFACTOR use `[]float64` as type
Rewriting the matrix lib to use a `[]float64` with "skip" multiplier
instead of using a `[][]float64` provided a 50%+ perf boost.

Most notably, this is because there are way less loops involved, and
generating a zero matrix is just about doing a
`make( Matrix, rows * cols + 2)` (and setting its dimensions values).